feat: Added manual JSON parsing in JsonUtils.java
Created helper methods to pull JSON values from JSON keys.
This was performed by parsing the string and removing
unwanted characters that could be presented to the user
through an instance of Sandwich. The UI was changed
to have a parent ConstraintLayout with an ImageView
and a child ScrollView. The ScrollView contained a
ConstraintLayout, which was used to present the information
in a clean manner.