fix: Replaced String parsing logic with JSONObject; Cleaned up UI
Previous JSON parsing was done directly on the JSON string instead of
converting the String to a JSONObject and then parsing the JSONObject
using its built-in methods.

The Sandwich image is now contained in the ScrollView and fixed an issue
where Place of Origin and Also Known As labels were being squished together
in landscape mode.

Moved String formatting logic for empty or null Strings to DetailActivity
instead of in JsonUtils.

Created a placeholder image for instances where an image is not available
from its URL.